+++ src/sll/include/sll/platform.h
@@ -502,7 +502,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_library_handle_t sll_platform_load_library
 
 /**
  * \flags check_output func
- * \name sll_platform_lookup_function
+ * \name sll_platform_lookup_symbol
  * \group platform
  * \subgroup platform-lib
  * \desc Docs!
@@ -510,7 +510,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_library_handle_t sll_platform_load_library
  * \arg const sll_char_t* nm
  * \ret void*
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_platform_lookup_function(sll_library_handle_t h,const sll_char_t* nm);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT void* sll_platform_lookup_symbol(sll_library_handle_t h,const sll_char_t* nm);