+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - *\[Darwin only\]* Fixed `SLL_API_MAX_FILE_PATH_LENGTH`
 - Fixed bit-shifting logic in `sll_load_assembly` and `sll_write_assembly`
 - Libraries loaded via `sys$load_library` can only be located in the `sys_lib` directory
+- Renamed `sll_platform_lookup_function` to `sll_platform_lookup_symbol`
 - Renamed API source files from `xxx_api.c` to `xxx.c`
 - Replaced `sll_platform_get_temporary_file_path` by `sll_temporary_file_path`